Basic support for LocalDate / Arrow Date types
Vector support for both 32bit epoch seconds, and 64bit epoch millisecond (DateDayVector and DateMilliVector).

Includes partial implementations (parity with timestamp) for DATE_TRUNC and EXTRACT.

I chose not to extend date/interval arithmetic at this time as it is a massive rabbit hole, agreed with @jarohen to do a date math pass at a later date.