handling lack of equi-join conditions (or any conditions, in fact) in inner/outer joins. impl n.b.:
moving the nil-row handling to IRelationMap's constructor. this ensured that the nil row didn't appear in its hashmap, and so wasn't inadvertently returned as a match...

as always with these kinds of fixes, it's worth considering at what point this becomes 'too clever'. I don't think we're there _just_ yet, but it's certainly something that we need to think about...