Postgres wire protocol server de-spiking.
Mostly rewritten using the spike code as the basis.

New connection negotiation, resource management.

Better handling of unexpected errors (connections are not held on to)

Parameters now work (though work needs to be done on types)

Docs and comments about why we are doing things, and to help understand lifecycles.