Add support for subqueries in qualified join clauses
Commit does this by creating nested dependant joins (apply)
where the result of any subquery is first joined onto the RHS and
then used as part of the join clause for the qualified join. This outer
join is responsible for brining in any parameters from the LHS so they
are accessable to the rest of the subtree.

Commit also fixes a bug in rewrite-equals-predicates-in-join-as-equi-join-map
which resulted in invalid join clause. As well as changing apply to no
longer pass params from its parent as this is no longer needed.